move tmp functions to libcore
Move the three temporary-file functions to libcore:

- mutt_buffer_mktemp_full()
- mutt_file_mkstemp_full()
- mutt_mktemp_full()

This includes their macro wrappers:

- mutt_buffer_mktemp()
- mutt_buffer_mktemp_pfx_sfx()
- mutt_file_mkstemp()
- mutt_mktemp()
- mutt_mktemp_pfx_sfx()

This makes them available to the unit tests, without adding dependencies
to libmutt.
