+++ muttlib.c
@@ -31,7 +31,6 @@
 #include "config.h"
 #include <ctype.h>
 #include <errno.h>
-#include <inttypes.h>
 #include <limits.h>
 #include <pwd.h>
 #include <stdbool.h>
@@ -454,63 +453,6 @@ bool mutt_is_text_part(struct Body *b)
   }
 
   return false;
-}
-
-/**
- * mutt_buffer_mktemp_full - Create a temporary file
- * @param buf    Buffer for result
- * @param prefix Prefix for filename
- * @param suffix Suffix for filename
- * @param src    Source file of caller
- * @param line   Source line number of caller
- */
-void mutt_buffer_mktemp_full(struct Buffer *buf, const char *prefix,
-                             const char *suffix, const char *src, int line)
-{
-  const char *const c_tmp_dir = cs_subset_path(NeoMutt->sub, "tmp_dir");
-  mutt_buffer_printf(buf, "%s/%s-%s-%d-%d-%" PRIu64 "%s%s", NONULL(c_tmp_dir),
-                     NONULL(prefix), NONULL(ShortHostname), (int) getuid(),
-                     (int) getpid(), mutt_rand64(), suffix ? "." : "", NONULL(suffix));
-
-  mutt_debug(LL_DEBUG3, "%s:%d: mutt_mktemp returns \"%s\"\n", src, line,
-             mutt_buffer_string(buf));
-  if (unlink(mutt_buffer_string(buf)) && (errno != ENOENT))
-  {
-    mutt_debug(LL_DEBUG1, "%s:%d: ERROR: unlink(\"%s\"): %s (errno %d)\n", src,
-               line, mutt_buffer_string(buf), strerror(errno), errno);
-  }
-}
-
-/**
- * mutt_mktemp_full - Create a temporary filename
- * @param buf    Buffer for result
- * @param buflen Length of buffer
- * @param prefix Prefix for filename
- * @param suffix Suffix for filename
- * @param src    Source file of caller
- * @param line   Source line number of caller
- *
- * @note This doesn't create the file, only the name
- */
-void mutt_mktemp_full(char *buf, size_t buflen, const char *prefix,
-                      const char *suffix, const char *src, int line)
-{
-  const char *const c_tmp_dir = cs_subset_path(NeoMutt->sub, "tmp_dir");
-  size_t n = snprintf(buf, buflen, "%s/%s-%s-%d-%d-%" PRIu64 "%s%s",
-                      NONULL(c_tmp_dir), NONULL(prefix), NONULL(ShortHostname),
-                      (int) getuid(), (int) getpid(), mutt_rand64(),
-                      suffix ? "." : "", NONULL(suffix));
-  if (n >= buflen)
-  {
-    mutt_debug(LL_DEBUG1, "%s:%d: ERROR: insufficient buffer space to hold temporary filename! buflen=%zu but need %zu\n",
-               src, line, buflen, n);
-  }
-  mutt_debug(LL_DEBUG3, "%s:%d: mutt_mktemp returns \"%s\"\n", src, line, buf);
-  if ((unlink(buf) != 0) && (errno != ENOENT))
-  {
-    mutt_debug(LL_DEBUG1, "%s:%d: ERROR: unlink(\"%s\"): %s (errno %d)\n", src,
-               line, buf, strerror(errno), errno);
-  }
 }
 
 /**
