+++ ncrypt/pgp.c
@@ -57,7 +57,6 @@
 #include "globals.h" // IWYU pragma: keep
 #include "handler.h"
 #include "hook.h"
-#include "muttlib.h"
 #include "pgpinvoke.h"
 #include "pgpkey.h"
 #include "pgpmicalg.h"
