+++ ncrypt/pgpmicalg.c
@@ -32,6 +32,7 @@
 #include <stdio.h>
 #include <string.h>
 #include "mutt/lib.h"
+#include "core/lib.h"
 #include "pgpmicalg.h"
 #include "handler.h"
 #include "pgppacket.h"
