+++ recvcmd.c
@@ -52,7 +52,6 @@
 #include "hdrline.h"
 #include "mutt_body.h"
 #include "mutt_logging.h"
-#include "muttlib.h"
 #include "protos.h"
 #ifdef ENABLE_NLS
 #include <libintl.h>
