+++ test/common.c
@@ -40,6 +40,7 @@ bool StartupComplete = true;
 char *HomeDir = NULL;
 int SigInt = 0;
 int SigWinch = 0;
+char *ShortHostname = "example";
 
 #define TEST_DIR "NEOMUTT_TEST_DIR"
 
@@ -161,11 +162,6 @@ struct Mailbox *get_current_mailbox(void)
 struct Menu *get_current_menu(void)
 {
   return NULL;
-}
-
-void mutt_mktemp_full(char *buf, size_t buflen, const char *prefix,
-                      const char *suffix, const char *src, int line)
-{
 }
 
 int mutt_do_pager(struct PagerView *pview, struct Email *e)
